+++ AM/src/SPSS.cc
@@ -173,7 +173,7 @@ void SPSS::cmptInitHessian(MatrixXd &d2f_dp2, const MatrixXd &dI0_dp){
 
 void SPSS::cmptInitHessian(MatrixXd &d2f_dp2, const MatrixXd &dI0_dp,
 	const MatrixXd &d2I0_dp2){
-	int p_size = d2f_dp2.rows();
+	int p_size = static_cast<int>(d2f_dp2.rows());
 
 	assert(d2f_dp2.cols() == p_size);
 	assert(d2I0_dp2.rows() == p_size * p_size);
@@ -204,7 +204,7 @@ void SPSS::cmptCurrHessian(MatrixXd &d2f_dp2, const MatrixXd &dIt_dp){
 }
 void SPSS::cmptCurrHessian(MatrixXd &d2f_dp2, const MatrixXd &dIt_dp,
 	const MatrixXd &d2It_dp2){
-	int p_size = d2f_dp2.rows();
+	int p_size = static_cast<int>(d2f_dp2.rows());
 	assert(d2f_dp2.cols() == p_size);
 	assert(d2It_dp2.rows() == p_size * p_size);
